pyproject.toml
[tool.poetry]
name = "aws-rag-bot"
version = "0.1.8"
description = ""
authors = ["danjamk <dan@risegardens.com>"]
license = "MIT"
readme = "README.md"
keywords = ["aws", "rag", "bedrock", "opensearch", "vector database"]
repository = "https://github.com/harvest2o-llc/aws-rag-bot"


[tool.poetry.dependencies]
python = ">=3.11,<3.12"
boto3 = "^1.34.51"
python-dotenv = "^1.0.1"
langchain = "^0.1.9"
langchain-openai = "^0.0.8"
opensearch-py = "^2.4.2"
requests = "^2.31.0"
requests-aws4auth = "^1.2.3"
langchain-google-genai = "^0.0.9"
transformers = "^4.38.1"
selenium = "^4.18.1"
unstructured = "^0.12.5"
torch = "^2.2.1"
anthropic = "^0.16.0"
sentence-transformers = "^2.5.0"
cohere = "^4.51"
docx2txt = "^0.8"



[tool.poetry.group.dev.dependencies]
ragas = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
